Move toxresult tests into their own file.
